+++ Roz.Identity.EntityFramework/RoleStore.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNet.Identity.EntityFramework;
 
 namespace Roz.Identity.EntityFramework
 {
-    public class RoleStore : RoleStore<Role, long, UserRole>
+    public class RoleStore : RoleStore<Role, int, UserRole>
     {
         public RoleStore(DbContext context)
             : base(context)
